+++ .eslintrc.json
@@ -11,6 +11,7 @@
         "sourceType": "module"
     },
     "rules": {
-        "semi": ["error", "always"]
+        "semi": ["error", "always"],
+        "no-import-assign": "off"
     }
 }
